// src/app/shared/model/ref-app-src-obj.model.ts
export class RefAppSrcObj {
    RefAppSrcId: number;
    AppSrcCode: string;
    AppSrcName: string;
    Descr: string;
    IsActive: boolean;
    MrAppSrcTypeCode: string;
    PeriodFrom: string;
    PeriodTo: string;
    MaxApvDt: Date;
    RowVersion: string;
    constructor() { this.RowVersion = "" }
  }
  